python_api/app/requirements.txt
flask==2.0.1
flask-cors
numpy
pandas
bs4
asyncio
matplotlib
requests
sqlalchemy
mysql-connector-python
python-dotenv
seaborn
sklearn
discord